lambdadb/requirements.txt
aws-cdk-lib==2.159.1
constructs>=10.0.0,<11.0.0

# -e .

aws-cdk.core
aws-cdk.aws_lambda
aws_cdk.aws_apigatewayv2
aws_cdk.aws_apigatewayv2_integrations